% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{oneway_1}
\alias{oneway_1}
\title{oneway_1}
\format{
\describe{
\item{group}{分组变量(空白对照组，去卵巢组，雌激素组)}
\item{tbar}{骨小梁面积百分比}
}
}
\description{
将27只3月龄雌性SD大鼠随机分为3组，每组9只。3组分别是空白对照组，去卵巢
组和雌激素组。90天后测量骨骼变化情况，用骨小梁面积百分比评价得到的结果
}
